requirements.txt
django>=1.11,<2
djangorestframework>3.6,<3.7
iso8601<0.2
